# 🚀 Add parallel parsing capabilities