--- a/docs/progress.md
+++ b/docs/progress.md
@@ -1,6 +1,5 @@
 # SharpParser.Core Implementation Progress
 
-## Project Status:  **COMPLETE**
 
 **Date**: October 5, 2024
 **Status**: All planned features implemented and documented